Mavenize example for OPENAM-3059
In this example, I have changed the package to
org.forgerock.openam.examples.

I also use demo:changeit instead of test:password for the user whose
auth is successful, since it prevents us from having to explain that
you need to create a test user.

Install it in the top-level realm, configure it, and then
http://openam.example.com:8080/openam/UI/Login?service=sample (logging
in as demo:changeit).